% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCombinedAgeKey.R
\name{BuildCombinedAgeKey}
\alias{BuildCombinedAgeKey}
\title{Create age-length key from combined acoustic survey and bottom trawl data.}
\usage{
BuildCombinedAgeKey(year = NULL, lake = NULL, target = c(1, 209, 210,
  223), user = username, password = password, dbname = dbname,
  schma = "RVCAT", outdir = NULL, species = c(106))
}
\arguments{
\item{year}{Numeric vector representing the year of data collection.}

\item{lake}{Numeric vector with the USGS-GLSC RVCAT numeric code for the
Great Lake in which data were collected.}

\item{target}{Numeric vector of the numeric code for the RVCAT target associated
with the desired data. Used to help limit the data to a desired study or survey.}

\item{user}{Character vector providing the username used to connect to
the database.}

\item{password}{Character vector providing the password used to connect to
the database.}

\item{schma}{Character vector providing the schema used to connect to
the database. Default is "RVCAT".}

\item{outdir}{Character vector representing the parent directory in which the
output directory where data should be saved is located. The values for lakec and
year are pasted together with this directory name to provide the output location
for files generated by the function. This folder must exist already.}

\item{species}{Numeric vector of the RVCAT species codes for the species
the user wishes to study.}
}
\value{
An age-length key formatted for EchoNet2Fish.
}
\description{
Create age-length key from combined acoustic survey and bottom trawl data.
}
\details{
This function is primarily a wrapper for fishmethods::alk() (Nelson 2018) combined
with an Oracle database query that is used to obtain the data.
}
\references{
Gary A. Nelson (2018). fishmethods: Fishery Science Methods and Models. R package version 1.11-0. https://CRAN.R-project.org/package=fishmethods
}
