% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVal.R
\name{colorVal}
\alias{colorVal}
\title{Colors Based on Values}
\usage{
colorVal(x, n = 100, low = 4/6, high = 0/6)
}
\arguments{
\item{x}{A numeric vector of values on which the colors are based.}

\item{n}{A numeric scalar giving the number of unique colors in the palette,
default 100.}

\item{low}{A numeric scalar giving the hue in [0, 1] at which the color palette
begins for low values, default 4/6 for blue.}

\item{high}{A numeric scalar giving the hue in [0, 1] at which the color palette
ends for high values, default 0 for red.}
}
\description{
Assign a range of unique colors based on values.
}
\examples{
rand <- rnorm(20)
plot(rand, rand, pch=16, col=colorVal(rand))

}
\seealso{
\code{\link{rainbow}}
}
