% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinecsv.R
\name{combinecsv}
\alias{combinecsv}
\title{Combine Several Comma Delimited Files into a Single Data Frame}
\usage{
combinecsv(myDir, addSource = TRUE, column1name = "Dummy_ID")
}
\arguments{
\item{myDir}{A character scalar naming the directory in which the csv files
are stored.  Should end in a forward slash, e.g., "C:/temp/".
All the csv files should have the same number of columns with the same
header row of column names.}

\item{addSource}{A logical scalar indicating whether a new column, named "source", should
be added to the data frame identifying the source file, default TRUE.}

\item{column1name}{A character scalar assigning a name to the first column in the data
frame (writing over whatever name is there already), default "Dummy_ID".}
}
\value{
A data frame with the information from all the csv files combined.
}
\description{
Combine all csv files in a given directory into a single data frame.
}
\details{
The \code{column1name} argument is needed to handle occasional problems
	 with byte order marks at the beginning of the csv files, which can result
	 in strange characters being added to the name of the first column.
	 See, for example, this
	 \href{http://stackoverflow.com/a/15399003/2140956}{link}.

  Any variable names starting with "X." and ending with a number
  are changed to ensure the number part of the name is rounded to the
  nearest whole number, e.g., "X.76.000000" is renamed "X.76".
}
