% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfPlot.R
\name{dfPlot}
\alias{dfPlot}
\title{Plot All Variables of a Data Frame}
\usage{
dfPlot(df, mcex = 1.2, cex = 0.8, ...)
}
\arguments{
\item{df}{A data frame to be plotted}

\item{mcex}{A numeric scalar giving the amount by which plot titles
will be magnified, default 1.2.}

\item{cex}{A numeric scalar giving the amount by which plotting text and symbols
will be magnified, default 0.8.}

\item{...}{Additional arguments to the \code{\link{par}} function.}
}
\description{
Plot all the variables of a data frame.
}
\details{
An individual graph is generated for each variable.  Bar plots are drawn
  for factors, characters with up to 50 unique values, and numerics
  with up to 10 unique values.  Otherwise finite values are plotted versus
  row number in \code{df}.  An empty plot is drawn for numerics with
  fewer than 2 finite values and characters with more than 50 unique values.
}
\examples{
mydat <- data.frame(a=c(1:5, 10:15, 6:9),
 b=as.factor(rep(c("cat", "dog", "frog", "cat", "dog"), 3)),
 c=rep(c("a", "b", "a", "a", "b"), 3),
 stringsAsFactors=FALSE)
dfPlot(mydat)

}
