% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfSmry.R
\name{dfSmry}
\alias{dfSmry}
\title{Brief Summary of a Data Frame}
\usage{
dfSmry(df)
}
\arguments{
\item{df}{A data frame to be summarized}
}
\value{
A matrix with a row for each variable (column) in \code{df} and three
  columns: \code{no.unique} the number of unique values,
  \code{no.entered} the number of rows in \code{df} with non-missing values,
  and \code{no.missing} the number of rows in \code{df} with missing values.
}
\description{
Brief summary of a data frame, number of unique and missing values.
}
\examples{
mydat <- data.frame(a=c(1, 1, NA, 2, 2),
 b=as.factor(c("cat", "dog", "frog", "", "dog")),
 c=c("a", "", "", "a", "b"),
 stringsAsFactors=FALSE)
dfSmry(mydat)

}
