% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTidy.R
\name{dfTidy}
\alias{dfTidy}
\title{Tidy Up a Data Frame}
\usage{
dfTidy(
  df,
  missNums = c(-9999, -999.9, -999, 999, 9999),
  missChars = c("NA", "NULL", ".")
)
}
\arguments{
\item{df}{A data frame to be tidied.}

\item{missNums}{A numeric vector of numbers representing missing values in \code{df},
default c(-9999, -999.9, -999, 999, 9999).}

\item{missChars}{A character vector of strings representing missing values in \code{df},
default c("NA", "NULL", ".", " ", "  ").}
}
\value{
A tidied version of the original data frame.
}
\description{
Tidy up a data frame with respect to missing values, internal quotes,
and white space.
}
\details{
Missing values in \code{df} are replaced with NA for numeric vectors
  and "" for character vectors.
  Internal quotes and leading and trailing white space
  in character vectors are removed.
  Attempt is made to convert any character vectors of numbers to numeric.
}
\examples{
mydat <- data.frame(a=c(1, 2, -999, 4, 5),
 b=as.factor(c("cat ", " dog", "frog", ".", " ant")),
 c=c("1.2", "NULL", "3.4", "5.6", "7.8"),
 stringsAsFactors=FALSE)
dfTidy(mydat)

}
