% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-data.R
\name{EchoEnv}
\alias{EchoEnv}
\title{EchoNet2Fish Package Local Environment}
\format{
An environment.
}
\source{
Post from Hadley Wickham to r-help on 2 Dec 2014
 \href{https://stat.ethz.ch/pipermail/r-help/2014-December/423847.html}{[link]}.
}
\description{
An environment local to the EchoNet2Fish package, used to hold objects
  outside of the individual package functions
}
