% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoNet2Fish-package.R
\docType{package}
\name{EchoNet2Fish}
\alias{EchoNet2Fish}
\title{Estimate Fish Abundance from Acoustic Echoes and Net Catch}
\description{
\pkg{EchoNet2Fish} estimates fish abundance from acoustic echoes and 
midwater trawl catch.
}
\details{
An example of how to use the functions in \pkg{EchoNet2Fish} is given in this
vignette
\href{https://rawgit.com/JVAdams/EchoNet2Fish/master/vignettes/Intro.html}{[link]}.
Use \code{\link{readAll}} to read in the acoustic and midwater trawl data.
Use \code{\link{exploreACMT}} and \code{\link{estimateACMT}} to 
  explore the data and estimate abundance.

\emph{U.S. Geological Survey} (USGS) Computer Program \pkg{EchoNet2Fish} 
  version 0.3.3.9000
Written by Jean V. Adams, USGS - Great Lakes Science Center
\href{http://www.glsc.usgs.gov/}{glsc.usgs.gov}, Ann Arbor, Michigan, USA.
Written in programming language R (R Core Team, 2018, www.R-project.org),
version 3.5.2 (2018-12-20).
Run on a PC with Intel(R) Core(TM) I7-7600U CPU, 2.90 GHz processor,
16.0 GB RAM, and Microsoft Windows 10 Enterprise operating system 2017.

Source code is available from Jean V. Adams on GitHub,
\href{https://github.com/JVAdams/EchoNet2Fish}{github.com/JVAdams/EchoNet2Fish},
\emph{jvadams (at) usgs (dot) gov}.

\emph{Disclaimer:}  Although this program has been used by the USGS,
no warranty, expressed or implied, is made by the USGS or the United States
Government as to the accuracy and functioning of the program and related
program material nor shall the fact of distribution constitute any such
warranty, and no responsibility is assumed by the USGS in connection
therewith.
}
