% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estNv.R
\name{estNv}
\alias{estNv}
\title{Estimate the Number of Targets in the Acoustic Beam}
\usage{
estNv(c = 1450, tau = 4e-04, psi, R, rhov)
}
\arguments{
\item{c}{Numeric scalar, sound speed in m/sec, default 1450.}

\item{tau}{Numeric scalar, pulse duration in sec, default 0.0004.}

\item{psi}{Numeric scalar, two-way equivalent beam angle in steradians.}

\item{R}{Numeric vector, range to target in m.}

\item{rhov}{Numeric vector, volumetric fish density in number per m^3, same length as \code{R},
typically the result of a call to \code{\link{estrhov}}.}
}
\value{
A numeric vector the same length as \code{R} containing the estimated
  number of targets in the acoustic beam.
}
\description{
Estimate the number of targets in the acoustic sampling volume.
}
\details{
The number of targets in the acoustic sampling volume
  is estimated according to Sawada et al. (1993) as
  Nv = \code{c} * \code{tau} * \code{psi} * \code{R}^2 * \code{rhov} / 2.
}
\examples{
mypsi <- 0.008
myrange <- c(10, 20, 30)
myrhov <- c(0.0001, 0.01, 0.0006)
estNv(psi=mypsi, R=myrange, rhov=myrhov)

}
\references{
Sawada, K., Furusawa, M., and Williamson, N.J. 1993.
  Conditions for the precise measurement of fish target strength in situ.
  Fisheries Science (Tokyo) 20:15-21.
}
\seealso{
\code{\link{estrhov}}
}
