% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estWeight.R
\name{estWEIGHT}
\alias{estWEIGHT}
\title{Estimate Fish WEIGHT From LENGTH}
\usage{
estWEIGHT(Lmm, LWa, LWb)
}
\arguments{
\item{Lmm}{A numeric vector, the total length of fish in mm.}

\item{LWa}{A numeric vector of the multiplicative parameter in a length-weight
relation, in log(g).  See details.}

\item{LWb}{A numeric vector of the exponential parameter in a length-weight
relation, in 1/log(mm).  See details.}
}
\value{
A numeric vector of fish weights (in g), the same length as \code{Lmm}.
}
\description{
Estimate fish weight from fish length.
}
\details{
The weight (in g) is calculated from the length-weight relation as
  Wg = \code{LWa} * \code{Lmm} ^ \code{LWb}.
}
\examples{
paDens(c(0.001, 0.01, 0.1), c(4e-5, 5e-6, 8e-9))

}
