% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figu.R
\name{figu}
\alias{figu}
\title{Add a Figure to an RTF Document}
\usage{
figu(
  ...,
  FIG = fig,
  rtf = doc,
  figid = "Figure ",
  fign = EchoEnv$figcount,
  boldt = TRUE,
  capunder = TRUE,
  w = NULL,
  h = NULL,
  rf = 300,
  newpage = "none",
  omi = c(1, 1, 1, 1)
)
}
\arguments{
\item{...}{One or more character scalars (separated by commas) of text to use for
the figure caption.}

\item{FIG}{A function to create a figure which will be added to the document,
default \code{fig}.}

\item{rtf}{An rtf object, default \code{doc}.}

\item{figid}{Character scalar of caption identifier, default "Figure ".}

\item{fign}{Numeric scalar of figure number to use in caption,
default \code{EchoEnv$figcount}.}

\item{boldt}{Logical scalar indicating if figure number should use bold font,
default TRUE.}

\item{capunder}{Logical scalar indicating if caption should appear
under the figure (TRUE, the default) or on top of the figure (FALSE).}

\item{w}{Numeric scalar width of figure in inches, default 6.5.}

\item{h}{Numeric scalar height of figure in inches, default 8.}

\item{rf}{Numeric scalar resolution of figure, default 300.}

\item{newpage}{Character scalar indicating if the figure should start on a new page in
the document "port" for a new portrait page, "land" for
a new landscape page, and "none" for no new page (the default).}

\item{omi}{Numeric vector, length 4, width of document page margins in inches
(bottom, left, top, right), default c(1, 1, 1, 1).}
}
\value{
A 1 is added to the numeric vector of length 1, \code{EchoEnv$figcount},
  stored in the working directory to keep track of the number of figures
  written to the rtf document, and label the captions accordingly.
}
\description{
Add a figure to an rtf (rich text format) document.
}
\details{
The figure and caption are written to the rtf file.
  The size of a new page is assumed to be 8.5 by 11 inches.
}
\references{
This is a copy of the \code{figu} function from the
  \href{https://github.com/JVAdams/GLFC}{[GLFC]} package.
}
\seealso{
\code{\link{startrtf}} for an example, \code{\link{heading}},
  \code{\link{para}}, \code{\link{tabl}},
  \code{\link{endrtf}},
  \code{\link[rtf]{RTF}}.
}
