% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLog.R
\name{groupLog}
\alias{groupLog}
\title{Assign Groups to Data on the Log Scale}
\usage{
groupLog(x, n, xR = range(x[x > 0 & !is.na(x) & is.finite(x)]))
}
\arguments{
\item{x}{Numeric vector of values to group.}

\item{n}{Integer scalar, the number of groups.}

\item{xR}{Numeric vector of length two, range of values used to define groups,
default is the range of finite, positive, non-missing \code{x} values.}
}
\description{
Assign groups to a collection of values, on the log-transformed scale.
}
\examples{
myx <- c(100, 3000, 3, 20, 8000, 1000, 400, 50, 7, 20000)
groupLog(myx, n=3)

}
