% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{Add a Heading to an RTF Document}
\usage{
heading(words, htype = 1, rtf = doc)
}
\arguments{
\item{words}{Character scalar text of heading to add to document.}

\item{htype}{Integer scalar heading type, 1=bold and font size 12, 2=bold and
font size 10, 3=italics and font size 10, default 1.}

\item{rtf}{An rtf object, default \code{doc}.}
}
\description{
Add a text heading to an rtf (rich text format) document.
}
\details{
The specified heading is written to the rtf file.
}
\references{
This is a copy of the \code{heading} function from the
  \href{https://github.com/JVAdams/GLFC}{[GLFC]} package.
}
\seealso{
\code{\link{startrtf}} for an example, \code{\link{para}},
  \code{\link{tabl}}, \code{\link{figu}},
  \code{\link{endrtf}},
  \code{\link[rtf]{RTF}}.
}
