% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histMulti.R
\name{histMulti}
\alias{histMulti}
\title{Multipanel Histogram of Values}
\usage{
histMulti(
  x,
  freq = rep(1, length(x)),
  bygroup,
  sug = sort(unique(bygroup)),
  xlab = "Values",
  ylab = "Frequency",
  barw = 5,
  samescale = TRUE,
  fillcol = "gray",
  rangecol = "red",
  mar = c(3, 3, 2, 1),
  oma = c(2, 2, 1, 1)
)
}
\arguments{
\item{x}{Numeric vector of values to be plotted.}

\item{freq}{Numeric vector of frequencies corresponding to \code{x}, default
a vector of 1s the  same length as \code{x}.}

\item{bygroup}{Vector, identifying the group membership of the values to be plotted.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(\code{bygroup})).}

\item{xlab}{Character scalar, label to assign to collection of x-axes.}

\item{ylab}{Character scalar, label to assign to collection of y-axes.}

\item{barw}{Numeric scalar, the width of the bars in units of \code{x}, default 5.}

\item{samescale}{Logical scalar, indicating if the same (TRUE, default) or different
(FALSE) lon/lat scales should be used for all panels.}

\item{fillcol}{A scalar, the color used to fill the histograms, default "gray".}

\item{rangecol}{A scalar, the color used to draw the lines denoting the range of values,
default "red".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around each plotted map plot, default
c(3, 3, 2, 1).}

\item{oma}{A numeric vector of length 4, the number of lines of the outer margin
c(bottom, left, top, right) around the entire collection of panels,
default c(2, 2, 1, 1).}
}
\description{
Multipanel histogram of values, one histogram for each group.
}
\examples{
myval <- c(23, 35, 47, 62)
myfreq <- c(5, 2, 3, 4)
mygroup <- c("A", "A", "B", "B")
histMulti(x=myval, freq=myfreq, bygroup=mygroup)
histMulti(x=myval, freq=myfreq, bygroup=mygroup, samescale=FALSE)

}
