% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon2utm.R
\name{latlon2utm}
\alias{latlon2utm}
\title{Convert Lon/Lat to UTM}
\usage{
latlon2utm(lon, lat, zone = NULL, quiet = TRUE)
}
\arguments{
\item{lon}{Numeric vector of longitudes.}

\item{lat}{Numeric vector of latitudes, same length as \code{lon}.}

\item{zone}{Numeric scalar of the UTM zone, default NULL.  See details.}

\item{quiet}{Logical scalar indicating whether to refrain from printing the determined
zone when \code{zone} is NULL, default TRUE.}
}
\value{
A data frame with two columns (easting and northing) and as many rows as
  the length of \code{lon} containing the converted UTM coordinates in
  meters.
}
\description{
Convert longitude and latitude coordinates to Universal Transverse
Mercator coordinates.
}
\details{
If \code{zone} is NULL, then the zone is determined from the medians
  of \code{lon} and \code{lat}.
}
\examples{
mylon <- c(-92.11, -76.47, -82.27, -83.42)
mylat <- c(46.76, 44.34, 44.76, 45.41)
latlon2utm(mylon, mylat, quiet=FALSE)
sapply(1:length(mylon), function(i)
  latlon2utm(mylon[i], mylat[i], quiet=FALSE))
latlon2utm(mylon, mylat, 18)

}
\references{
Based on a function posted by Stanislav on 13 May 2015 on stackoverflow
  \href{http://stackoverflow.com/a/30225804/2140956}{[link]}.
}
\seealso{
\code{\link{lon2utmZone}}
}
