% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon2utmZone.R
\name{lon2utmZone}
\alias{lon2utmZone}
\title{Determine UTM Zone from Longitude}
\usage{
lon2utmZone(lon, lat = NULL)
}
\arguments{
\item{lon}{Numeric vector of longitudes.}

\item{lat}{Numeric vector of latitudes, same length as \code{lon}, default NULL.}
}
\value{
A numeric vector giving the UTM zones (integers) containing the \code{lon}.
  If \code{lat} is not NULL, then zones are corrected to account for
  the special cases of Svalbard and parts of Norway.
}
\description{
Determine the Universal Transverse Mercator zone that contains the given
longitude coordinates.
}
\examples{
mylon <- c(-82.27, NA, -122.44, 21.97, 21.97)
mylat <- c(44.76, 37.77, 37.77, NA, 78.64)
lon2utmZone(mylon)
lon2utmZone(mylon, mylat)

}
\references{
Based on a function posted by Josh O'Brien on 8 Feb 2012 on stackoverflow
  \href{http://stackoverflow.com/a/9188972/2140956}{[link]}.
  and code for exceptions posted by wittrup on 13 Sep 2013 on stackoverflow
  \href{http://stackoverflow.com/a/18785421/2140956}{[link]}.
}
