% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapACstrata.R
\name{mapACstrata}
\alias{mapACstrata}
\title{Multipanel Map of Locations}
\usage{
mapACstrata(bygroup, sug = sort(unique(bygroup)), lon, lat,
  rlon = range(lon, na.rm = TRUE) + 0.1 * c(-1, 1), rlat = range(lat, na.rm
  = TRUE) + 0.1 * c(-1, 1), IDcol = NULL, mapcol = "gray", mar = c(0, 0,
  0, 0))
}
\arguments{
\item{bygroup}{Vector, identifying the group membership of the locations to be mapped.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(\code{bygroup})).}

\item{lon}{A numeric vector of longitudes in decimal degrees.
Same length as\code{lat}.}

\item{lat}{A numeric vector of latitudes in decimal degrees.}

\item{rlon}{A numeric vector of length 2, range of longitudes to map,
in decimal degrees, default
range(\code{lon}, na.rm=TRUE) + 0.1*c(-1, 1).}

\item{rlat}{A numeric vector of length 2, range of latitudes to map,
in decimal degrees, default
range(\code{lat}, na.rm=TRUE) + 0.1*c(-1, 1).}

\item{IDcol}{A vector, the color used to map locations, same length as \code{bygroup}.
If NULL, the default, a range of colors will be assigned automatically.}

\item{mapcol}{A scalar, the color used to draw the map lines (e.g., lake boundary),
default "gray".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around the plotted map plot, default
c(0, 0, 0, 0).}
}
\description{
Multipanel map of locations, one map for each group.
}
\examples{
\dontrun{
 mygroup <- c(1, 1, 1, 2, 2, 2, 3, 3, 3)
 mylon <- -c(81.1, 81.2, 80.5, 83, 82.2, 82.7, 82.7, 82, 82.2)
 mylat <- c(45.7, 45.4, 45, 45.5, 45.4, 45, 44.4, 44.4, 43.9)
 mapACstrata(bygroup=mygroup, lon=mylon, lat=mylat)
}

}

