% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBy2Groups.R
\name{mapBy2Groups}
\alias{mapBy2Groups}
\title{Multipanel Map of Locations, with Gaps}
\usage{
mapBy2Groups(
  df,
  lon,
  lat,
  rlon = range(lon, na.rm = TRUE),
  rlat = range(lat, na.rm = TRUE),
  nrows,
  nsymbols = 7,
  mar = c(0, 0, 3, 0)
)
}
\arguments{
\item{df}{Data frame of values to plot, column names should identify membership
in two groups, separated by a period, "group1.group2".}

\item{lon}{Numeric vector, longitudes of locations to map,
in decimal degrees, same length as \code{MTgroup}.}

\item{lat}{Numeric vector, latitudes of locations to map,
in decimal degrees, same length as \code{MTgroup}.}

\item{rlon}{A numeric vector of length 2, range of longitudes to map,
in decimal degrees, default range(\code{lon}, na.rm=TRUE).}

\item{rlat}{A numeric vector of length 2, range of latitudes to map,
in decimal degrees, default range(\code{lat}, na.rm=TRUE).}

\item{nrows}{A numeric scalar, the number of rows of plot panels in the page.}

\item{nsymbols}{An integer scalar, between 1 and 7 inclusive,
the number of different symbol sizes/colors to plot,
default 7.}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around each plotted map plot, default
c(0, 0, 3, 0).}
}
\description{
Multipanel map of locations, arranged according to membership in the first
of two different groupings, with gaps (empty plot panels)
between the first groups.
}
\examples{
\dontrun{
 mydf <- as.data.frame(matrix(rlnorm(35, 5), nrow=7, dimnames=list(NULL,
   c("Dog.large", "Dog.small", "Cat.large", "Cat.small", "Fish.all"))))
 mylon <- -c(83, 83, 80.4, 81)
 mylat <- c(45.4, 44.5, 45, 45.5)
 mapBy2Groups(df=mydf, lon=mylon, lat=mylat, nrows=3)
 }

}
