% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapByGroup.R
\name{mapByGroup}
\alias{mapByGroup}
\title{Draw a Map using Different Colored Symbols for Groups}
\usage{
mapByGroup(
  bygroup,
  sug = sort(unique(bygroup)),
  addMean = TRUE,
  lon,
  lat,
  rlon = range(lon, na.rm = TRUE),
  rlat = range(lat, na.rm = TRUE),
  cushion = 0.1,
  colorz = NULL,
  pch = 1,
  cex = 1.5,
  mapcol = "gray",
  mar = c(0, 0, 2, 0)
)
}
\arguments{
\item{bygroup}{Vector, identifying the group membership of the locations to be mapped.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(\code{bygroup})).}

\item{addMean}{Logical scalar indicating if group names should be added to plot at the
mean location of each bygroup, default TRUE.}

\item{lon}{A numeric vector of longitudes in decimal degrees.}

\item{lat}{A numeric vector of latitudes in decimal degrees.
Same length as \code{lon}.}

\item{rlon}{A numeric vector of length 2, range of longitudes to map,
in decimal degrees, default is the range of \code{lon}.}

\item{rlat}{A numeric vector of length 2, range of latitudes to map,
in decimal degrees, default is the range of \code{lat}.}

\item{cushion}{A numeric scalar indicating the amount of cushion to add to the \code{rlon}
and \code{rlat} ranges in decimal degrees, default 0.1.}

\item{colorz}{A vector of colors to use, either of length 1 or the same length as
\code{lon}.  If NULL, the default, a range of colors will be assigned
automatically.}

\item{pch}{A vector of plotting characters or symbols, either of length
1 or the same length as \code{lon}, default 1.
See \code{\link[graphics]{points}}.}

\item{cex}{A numeric vector giving the amount by which plotting characters and
symbols should be scaled relative to the default, either of length
1 or the same length as \code{lon}, default 1.5.}

\item{mapcol}{A scalar, the color used to draw the map lines (e.g., lake boundary),
default "gray".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around the plotted map plot, default
c(0, 0, 0, 0).}
}
\description{
Draw a map using different colored symbols for different groups.
}
\examples{
\dontrun{
 mygroup <- c(1, 1, 1, 2, 2, 2, 3, 3, 3)
 mylon <- -c(81.1, 81.2, 80.5, 83, 82.2, 82.7, 82.7, 82, 82.2)
 mylat <- c(45.7, 45.4, 45, 45.5, 45.4, 45, 44.4, 44.4, 43.9)
 mapByGroup(bygroup=mygroup, lon=mylon, lat=mylat, cushion=0.7)
}

}
