% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMulti.R
\name{mapMulti}
\alias{mapMulti}
\title{Multipanel Map of Locations}
\usage{
mapMulti(bygroup, sug = sort(unique(bygroup)), ID, boldID = NULL,
  short = TRUE, lon, lat, rlon = range(lon, na.rm = TRUE),
  rlat = range(lat, na.rm = TRUE), IDcol = NULL, mapcol = "gray",
  boxcol = "gray", misscol = "brown", misstext = " - No sites",
  mar = c(0, 0, 2.5, 0))
}
\arguments{
\item{bygroup}{Vector, identifying the group membership of the locations to be mapped.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(\code{bygroup})).}

\item{ID}{Vector, text to print on the map at each location,
same length as \code{bygroup}.}

\item{boldID}{Vector, subset of unique \code{ID}s to be emphasized on the map
with large bold font.}

\item{short}{Logical scalar, indicating aspect of map area.  If TRUE, the default,
the mapped area is assumed to be wider (longitudinally) than tall.
Used to better arrange multiple maps on a single page.}

\item{lon}{A numeric vector of longitudes in decimal degrees.
Same length as\code{lat}.}

\item{lat}{A numeric vector of latitudes in decimal degrees.}

\item{rlon}{A numeric vector of length 2, range of longitudes to map,
in decimal degrees, default range(\code{lon}, na.rm=TRUE).}

\item{rlat}{A numeric vector of length 2, range of latitudes to map,
in decimal degrees, default range(\code{lat}, na.rm=TRUE).}

\item{IDcol}{A vector, the color used to map locations, same length as \code{bygroup}.
If NULL, the default, a range of colors will be assigned automatically.}

\item{mapcol}{A scalar, the color used to draw the map lines (e.g., lake boundary),
default "gray".}

\item{boxcol}{A scalar, the color used to draw the box around the map, default "gray".}

\item{misscol}{A scalar, the color used to label maps with no locations in the
given bygroup, default "brown".}

\item{misstext}{A character scalar, the text used to label maps with no locations in the
given bygroup, default " - No sites".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around each plotted map plot, default
c(0, 0, 2.5, 0).}
}
\description{
Multipanel map of locations, one map for each group.
}
\examples{
\dontrun{
 mygroup <- c(1, 1, 1, 2, 2, 2, 3, 3, 3)
 myID <- LETTERS[1:9]
 mylon <- rnorm(9, mean=-82)
 mylat <- rnorm(9, mean=45)
 mapMulti(bygroup=mygroup, ID=myID, boldID=c("G", "A"), lon=mylon, lat=mylat)
 mapMulti(bygroup=mygroup, sug=1:4, ID=myID, lon=mylon, lat=mylat)
}

}

