% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapText.R
\name{mapText}
\alias{mapText}
\title{Add Text to a Map}
\usage{
mapText(lat, long, group, cex = 1.5, ...)
}
\arguments{
\item{lat}{A numeric vector of latitudes in decimal degrees.}

\item{long}{A numeric vector of longitudes in decimal degrees.
Same length as\code{lat}.}

\item{group}{A character or numeric vector of group identifiers,
the same length as \code{lat} and \code{long}.}

\item{cex}{A numeric scalar giving the amount by which plotting characters
should be scaled relative to the default, default 1.5.}

\item{...}{Additional arguments to \code{\link[graphics]{text}}.}
}
\description{
Add identifying text to a map based on a single grouping variable.
}
\details{
Group identifiers are added to a map, typically created with
	 \code{\link{mapSymbols}}, at the midpoint of the range of each groups'
	 latitudes and longitudes
}
\examples{
\dontrun{
 latitude <- c(43.25, 45.73, 45.71, 44.84)
 longitude <- c(-82.30, -80.85, -84.03, -80.39)
 basincode <- c(1, 2, 1, 2)
 basin <- c("Main", "GBay", "Main", "GBay")
 mapSymbols(lat=latitude, long=longitude, colorz=basincode+3,
  pch=16, xla=0.4)
 mapText(lat=latitude, long=longitude, group=basin)
}
}

