% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint.R
\name{midpoint}
\alias{midpoint}
\title{Midpoint Between the Minimum and the Maximum}
\usage{
midpoint(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar representing the midpoint between the minimum and the
  maximum of \code{x}, ignoring missing values.
}
\description{
Calculate the midpoint between the minimum and the maximum of a vector.
}
\examples{
midpoint(c(10:20, 90))
}

