% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paDens.R
\name{paDens}
\alias{paDens}
\title{Density Per Unit Area}
\usage{
paDens(sigmabs, ABC, hectare = TRUE)
}
\arguments{
\item{sigmabs}{A numeric vector, the backscattering cross section, sigma_bs in m^2.}

\item{ABC}{A numeric vector, area backscattering coefficient, s_a (unitless),
the same length as \code{sigmabs}.}

\item{hectare}{A logical scalar, indicating if the density should be calculated as the
number per hectare (the default, TRUE), or the number per m^2 (FALSE).}
}
\value{
A numeric vector of fish densities, p_a in number per unit area,
  the same length as \code{sigmabs}.
}
\description{
Calculate the density of fish in number per unit area
from the backscattering cross
section, sigma_bs in m^2, and the area backscattering coefficient,
s_a (unitless).
}
\details{
For density in number per m^2, p_a = s_a / sigma_bs.
}
\examples{
paDens(c(0.001, 0.01, 0.1), c(4e-5, 5e-6, 8e-9))

}
