% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{para}
\alias{para}
\title{Add a Paragraph to an RTF Document}
\usage{
para(..., rtf = doc, bold = FALSE, italic = FALSE)
}
\arguments{
\item{...}{One or more character scalars (separated by commas) of text to add to
document as a single paragraph.}

\item{rtf}{An rtf object, default \code{doc}.}

\item{bold}{Logical scalar indicating if paragraph should use bold font, default FALSE.}

\item{italic}{Logical scalar indicating if paragraph should use italic font,
default FALSE.}
}
\description{
Add a paragraph to an rtf (rich text format) document.
}
\details{
The specified heading is written to the rtf file.
}
\references{
This is a copy of the \code{para} function from the
  \href{https://github.com/JVAdams/GLFC}{[GLFC]} package.
}
\seealso{
\code{\link{startrtf}} for an example, \code{\link{heading}},
  \code{\link{tabl}}, \code{\link{figu}},
  \code{\link{endrtf}},
  \code{\link[rtf]{RTF}}.
}
