% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACMTslice.R
\name{plotACMTslice}
\alias{plotACMTslice}
\title{Multipanel Plot of AC and MT Samples by Slice}
\usage{
plotACMTslice(
  MTgroup,
  ACgroup,
  sug = sort(unique(c(MTgroup, ACgroup))),
  MTbd,
  ACbd,
  MTwd,
  ACwd,
  Gcol = NULL,
  boxcol = "gray",
  axlabelcol = "gray",
  axtitlecol = "darkgray",
  mar = c(0, 0, 3, 3),
  oma = c(1.5, 2, 2.5, 2)
)
}
\arguments{
\item{MTgroup}{Vector, identifying the group membership (typically, the slice)
of the midwater trawl tow locations to be mapped.}

\item{ACgroup}{Vector, identifying the group membership (typically, the slice)
of the acoustic transect locations to be mapped.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(c(\code{MTgroup}, \code{ACgroup}))).}

\item{MTbd}{Numeric vector of bottom depths sampled by midwater trawl tows,
same length as \code{MTgroup}.}

\item{ACbd}{Numeric vector of bottom depths sampled by acoustic transects,
same length as \code{ACgroup}.}

\item{MTwd}{Numeric vector of water depths sampled by midwater trawl tows,
same length as \code{MTgroup}.}

\item{ACwd}{Numeric vector of water depths sampled by acoustic transects,
same length as \code{ACgroup}.}

\item{Gcol}{A vector, the color used for plot symbols,
same length as \code{sug}.
If NULL, the default, colors will be assigned automatically.}

\item{boxcol}{A scalar, the color of the three- to five-sided box around the scatterplot,
default "gray".}

\item{axlabelcol}{A scalar, the color of the axis labels, default "gray".}

\item{axtitlecol}{A scalar, the color of the axis titles, default "darkgray".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around each plotted map plot, default
c(0, 0, 2.5, 0).}

\item{oma}{A numeric vector of length 4, the number of lines of outer margin
c(bottom, left, top, right) around the multipanel plots, default
c(1.5, 2, 1.5, 2).}
}
\description{
Multipanel plot of acoustics and midwater trawl samples by slice.
}
\examples{
# acoustic transects
agroup <- c("Main", "Bay")[c(1, 1, 1, 1, 2, 2, 2)]
abd <- c(50, 100, 50, 100, 50, 100, 50)
awd <- c(10, 10, 20, 20, 10, 10, 20)

# midwater trawls
mgroup <- c("Main", "Bay")
mbd <- c(30, 80)
mwd <- c(15, 60)

plotACMTslice(MTgroup=mgroup, ACgroup=agroup, MTbd=mbd, ACbd=abd,
  MTwd=mwd, ACwd=awd)

}
