% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpecies.R
\name{plotSPECIES}
\alias{plotSPECIES}
\title{Plot Boxplots of Continuous Data by SPECIES}
\usage{
plotSPECIES(y, ylabb, x = SPECIES, uniqSp = sort(unique(x)))
}
\arguments{
\item{y}{A numeric vector of data to plot against species, the same length as
\code{x}.}

\item{ylabb}{A character scalar of the y-axis label.}

\item{x}{A vector (can be numeric, character, or factor) identifying species,
default \code{SPECIES}.}

\item{uniqSp}{A vector of unique species, a subset of \code{x}, default
\code{sort(unique(x))}.}
}
\description{
Boxplots of species versus continuous data on the square root scale.
}
