% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAll.R
\name{readAll}
\alias{readAll}
\title{Read in All the Data Corresponding to a Specified Reference Row}
\usage{
readAll(
  refdir,
  keyvals,
  keyvars = c("LAKE", "YEAR"),
  rdat = "ACMT",
  refcsv = "Reference"
)
}
\arguments{
\item{refdir}{A character scalar giving the directory in which the reference csv file,
\code{ref}, is located.}

\item{keyvals}{A vector giving the values of \code{keyvars} to identify a single row of
the reference csv file, \code{ref}.}

\item{keyvars}{A character vector giving the names of columns in the reference csv file,
\code{ref}, that will be used to identify a single row, default
c("LAKE", "YEAR").}

\item{rdat}{A character scalar giving the name of the RData file that will be saved
containing all of the objects corresponding to the selected row of
the reference csv file, \code{ref}, default "ACMT".}

\item{refcsv}{A character scalar giving the name of the reference csv file,
\code{ref}, default "Reference".  The csv file must have the following
10 columns:
\itemize{
  \item \code{subdir} = a subdirectory of \code{refdir} containing
  all the other subdirectories and files,
  \item \code{svsubdir} = the Sv subdirectory,
  \item \code{tssubdir} = the TS subdirectory,
  \item \code{optropf} = the midwater trawl operations file,
  \item \code{trcatchf} = the midwater trawl catch file,
  \item \code{trlff} = the midwater trawl lengths file,
  \item \code{keysp1} = the species code for \code{keyfile1},
  \item \code{keyfile1} = the age-length csv file for species\code{keysp1},
  \item \code{keysp2} = the species code for \code{keyfile2},
  \item \code{keyfile2} = the age-length csv file for species\code{keysp2}.
}
There should also be additional columns for \code{keyvars}.}
}
\value{
A character scalar giving the path of the subdirectory
  containing all of inputs and outputs for the given \code{keyvars} and
  \code{keyvals}.
}
\description{
Read in all the acoustic and midwater trawl data corresponding to the
specified row or a reference csv file.  Save the data as objects in
an RData file.
}
\details{
The acoustic and midwater trawl data corresponding to the selected row
  of the reference csv file are read in and saved as objects in the
  specified RData file, in \code{subdir}.  Objects include
  the scalar
    \code{rdat};
  the vectors
    \code{keyvals} and \code{keyvars};
  and the data frames
    \code{inputs} (the selected row from the reference data);
    \code{sv} and \code{ts} (acoustic data);
    \code{optrop}, \code{trcatch}, and \code{trlf} (midwater trawl data);
    \code{key1} and \code{key2} (age-length keys, if specified).
}
