% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSVTS.R
\name{readSVTS}
\alias{readSVTS}
\title{Read in Sv or TS Data from Echoview}
\usage{
readSVTS(
  svtsdir,
  oldname = NULL,
  newname = NULL,
  elimMiss = NULL,
  datevars = c("Date_S", "Date_E", "Date_M"),
  addyear = TRUE,
  tidy = TRUE
)
}
\arguments{
\item{svtsdir}{A character scalar giving the directory in which the csv files are located.}

\item{oldname}{A character vector giving the original names of variables to be renamed,
default NULL.}

\item{newname}{A character vector giving the new names of variables to be renamed,
default NULL.}

\item{elimMiss}{A character vector giving the names of variables which must have
nonmissing values.  All rows with a missing value for any of these
variables will be eliminated, default NULL.}

\item{datevars}{A character vector giving the names of variables with date information
stored as YYYYMMDD, default c("Date_S", "Date_E", "Date_M").
These will be converted to R dates.
If NA, no date coversion will be carried out.}

\item{addyear}{A logical scalar indicating if a \code{year} variable should be added
to the data frame, based on the first variable named in
\code{datevars}, default TRUE.}

\item{tidy}{A logical scalar indicating if the data frame should be tidied using
\code{\link{dfTidy}}, default TRUE.}
}
\value{
A data frame with the all rows of the csv files in \code{svtsdir} combined.
}
\description{
Read in Sv (volume backscattering strength) or TS (single target frequencies)
data from Echoview csv files and combine into a single data frame.
}
