% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma2TS.R
\name{sigma2TS}
\alias{sigma2TS}
\title{Convert Backscattering Cross Section to Target Strength}
\usage{
sigma2TS(sigmabs)
}
\arguments{
\item{sigmabs}{A numeric vector of backscattering cross sections, sigma_bs, in m^2.}
}
\value{
A numeric vector of target strengths, TS, in dB, same
  length as \code{sigmabs}.
}
\description{
Convert backscattering cross section (sigma_bs in m^2) to
target strength (TS in dB).
}
\details{
TS = 10*log10(sigma_bs).
}
\examples{
sigma2TS(c(1e-03, 1e-04, 1e-05, 1e-06))

}
\seealso{
\code{\link{TS2sigma}}
}
