% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliceCat.R
\name{sliceCat}
\alias{sliceCat}
\title{Categorize Observations as Slices for Matching Acoustic Densities and Trawl
Catches}
\usage{
sliceCat(sliceDef, fdp = NULL, bdp = NULL, lat = NULL, reg = NULL)
}
\arguments{
\item{sliceDef}{A list of at least two named sub-lists defining the slices into which
observations will be classified.  Each sub-list contains one or more
named numeric vectors of length two, identifying the parameter
(the name of the vector) and the range of values that contribute
to the slice definition.
Each interval is closed on the left and open on the right (see Details).
The name of each sub-list is the name of the slice to be assigned.
See Examples.}

\item{fdp}{A numeric vector of fishing depths (the distance from the surface of the
water to the depth of a fish in the water) corresponding to the
observations which are to be categorized into slices.  Only necessary if
required by \code{sliceDef}, default NULL.
default}

\item{bdp}{A numeric vector of bottom depths (the distance from the surface of the
water to the substrate) corresponding to the
observations which are to be categorized into slices.  Only necessary if
required by \code{sliceDef}, default NULL.}

\item{lat}{A numeric vector of latitudes corresponding to the
observations which are to be categorized into slices.  Only necessary if
required by \code{sliceDef}, default NULL.}

\item{reg}{A character vector of regions corresponding to the
observations which are to be categorized into slices.  Only necessary if
required by \code{sliceDef}, default NULL.}
}
\value{
A character vector the same length as the observations variables
  (\code{fdp}, \code{bdp}, \code{lat}, \code{reg}), identifying the slice to
  which each observation belongs.
}
\description{
Categorize observations as spatial slices for matching fish densities
estimated from acoustic transects and speciec compositions estimated from
midwater trawl catches.
}
\details{
Each interval of sliceDef is closed on the left and open on the right.
In other words, if you assign an interval of fdp=c(10, 20), observations
>= 10 and < 20 will be considered for inclusion in that slice.

All observation variables (\code{fdp}, \code{bdp}, \code{lat}, \code{lat}),
if not NULL, must be the same length.
}
\examples{
myslicedef <- list(
  epiNear = list( fdp=c(0,   4), bdp=c(0,   6) ),
  epiOff =  list( fdp=c(0,   4), bdp=c(6, Inf) ),
  hypo =    list( fdp=c(4, Inf) )
)
fishingD <- 1:7
bottomD <- c(2, 10, 4, 12, 6, 14, 8)
slice <- sliceCat(myslicedef, fdp=fishingD, bdp=bottomD)
data.frame(fishingD, bottomD, slice)
}

