% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratClust.R
\name{stratClust}
\alias{stratClust}
\title{Estimation from a Stratified Survey Design}
\usage{
stratClust(
  elementdf,
  stratum,
  cluster,
  response,
  sizedf,
  stratum2 = stratum,
  size
)
}
\arguments{
\item{elementdf}{A data frame with one row for each sample unit (element),
containing variables identifying the stratum, cluster, and
response value for each element.}

\item{stratum}{A character scalar giving the name of the variable in \code{elementdf}
that identifies the strata.}

\item{cluster}{A character scalar giving the name of the variable in \code{elementdf}
that identifies the clusters.}

\item{response}{A character scalar giving the name of the variable in \code{elementdf}
that contains the response values.}

\item{sizedf}{A data frame with one row for each stratum.
It should contain variables identifying the stratum and the
size of each stratum.}

\item{stratum2}{A character scalar giving the name of the variable in \code{sizedf}
that identifies the strata, default is \code{stratum},
the same name as in \code{elementdf}.}

\item{size}{A character scalar giving the name of the variable in \code{sizedf}
that contains the sizes of the strata.}
}
\value{
A list of three data frames (tibbles, actually).
  \code{Cluster} has a row for each cluster with four variables:
  \itemize{
    \item \code{h} = stratum (may be character, factor, or numeric)
    \item \code{i} = cluster (may be character, factor, or numeric)
    \item \code{m_hi} = number of elements in stratum h, cluster i (numeric)
    \item \code{y_hi} = sum of the response for stratum h, cluster i (numeric)
  }
  \code{Stratum} has a row for each stratum with eight variables:
  \itemize{
    \item \code{h} = stratum (may be a character, factor, or numeric)
    \item \code{m_h} = number of elements in stratum h (numeric)
    \item \code{y_h} = sum of response in stratum h (numeric)
    \item \code{n_h} = number of clusters in stratum h (numeric)
    \item \code{ybar_h} = mean response for stratum h (numeric)
    \item \code{s_ybar_h} = standard deviation of ybar_h (numeric)
    \item \code{A_h} = size of stratum h (numeric)
    \item \code{W_h} = relative size of stratum h (numeric)
  }
  \code{Population} has one row with five numeric variables:
  \itemize{
    \item \code{A} = total size of all strata in population
    \item \code{ybar_str} = mean response for population per unit of size
    \item \code{s_ybar_str} = standard deviation of ybar_str
    \item \code{ytot_str} = total response for population
    \item \code{s_ytot_str} = standard deviation of ytot_str
  }
}
\description{
Estimate the population response from a stratified survey design in which
single stage cluster samples were taken from each stratum, and the size of
each stratum is known.
}
\examples{
# Example data from a stratified survey design in which
# single stage cluster sampling is used in each stratum.
# Strata are areal regions of a lake, and the response are counts of fish.
counts <- data.frame(
 Stratum=rep(c("A", "B", "C"), c(5, 8, 8)),
 Cluster=rep(1:8, c(3, 2, 3, 2, 3, 2, 3, 3)),
 Element=c(1, 2, 3, 1, 2, 1, 2, 3, 1, 2, 1, 2, 3, 1, 2, 1, 2, 3, 1, 2, 3),
 Count = c(5:1, 6:21)
)
# Surface area (in hectares) corresponding to each lake stratum.
areas <- data.frame(
 Stratum=c("A", "B", "C"),
 A_h=c(10, 20, 40)
)
stratClust(elementdf=counts, stratum="Stratum", cluster="Cluster",
 response="Count", sizedf=areas, size="A_h")

}
\references{
Cochran, W.G. 1977. \href{https://archive.org/details/Cochran1977SamplingTechniques_201703}{[Sampling Techniques]}. Wiley, New York.
Scheaffer, R.L., Mendenhall III, W., Ott, R.L., Gerow, K., 2011. Elementary Survey Sampling, 7th ed. Brooks/Cole, Boston, MA.
}
