% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS2sigma.R
\name{TS2sigma}
\alias{TS2sigma}
\title{Convert Target Strength to Backscattering Cross Section}
\usage{
TS2sigma(TS)
}
\arguments{
\item{TS}{A numeric vector of target strengths, TS, in dB.}
}
\value{
A numeric vector of backscattering cross sections, sigma_bs, in m^2, same
  length as \code{TS}.
}
\description{
Convert target strength (TS in dB) to backscattering cross section
(sigma_bs in m^2), <sigma_bs>.
}
\details{
sigma_bs = 10^(TS/10).
}
\examples{
TS2sigma(c(-30, -40, -50, -60))

}
\seealso{
\code{\link{sigma2TS}}
}
