% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtLenCoef.R
\name{WtLenCoef}
\alias{WtLenCoef}
\title{Calculate Nonlinear Estimates of WEIGHT-LENGTH Coefficients}
\usage{
WtLenCoef(
  lake = 2,
  year = 2019,
  species = 106,
  target = c(1, 209, 210, 223, 252),
  user = username,
  password = password,
  dbname = dbname,
  schma = "RVCAT"
)
}
\arguments{
\item{lake}{Numeric vector to select lake from RVCAT lake codes.}

\item{year}{Numeric vector to select year from RVCAT.}

\item{species}{Numeric vector to select species.}

\item{target}{Numeric vector to select target.}

\item{user}{RVCAT username.}

\item{password}{RVCAT password.}

\item{dbname}{Database name.}

\item{schma}{Schema to use, set as RVCAT.}
}
\value{
Coefficients for weight-length equation.
}
\description{
Calculate Nonlinear Estimates of WEIGHT-LENGTH Coefficients
}
