% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchogramCurtain.R
\name{EchogramCurtain}
\alias{EchogramCurtain}
\title{Create echogram curtain plot from interval-layer data.}
\usage{
EchogramCurtain(
  the.dir,
  d.units,
  spgrp = c("X106.L0", "X106.L100", "X109.L0", "X109.L90", "X204.L0", "X204.L120"),
  QuantileFilter = FALSE,
  Quantile = 0.99
)
}
\arguments{
\item{the.dir}{Character, the directory containing the data that will be used to make the
plots.}

\item{d.units}{Character, either "nph" or "gph".}

\item{spgrp}{Character vector listing the species groups that should be plotted.
Must be entered in the form "X106.L100", for example, for alewife
(species code = 106) that are >= 100 mm long. These names are
in the form generated by the estimateLake() function.}

\item{QuantileFilter}{Logical scalar indicating whether or not the user wants to plot all
values as they are or if the user wants to assign the values above
a particulare percentile the value of the chosen percentile. The intent
of this parameter is to allow one to choose to use the plots to find
acoustic cells that might have unreasonable density and need to be checked
or, instead to plot distribution after confirming that any really high
values appear to be real. This allows one to reduce the influence of a
few very high values on distribution patterns.}

\item{Percentile}{Numeric scalar. The data values that are equal to or greater than
this percentile will be assigned the value of this percentile.}
}
\value{
This function generates curtain plots for each species group
listed and every acoustic transect. The transects are facets,
the vertical axis is depth, and the horizontal axis is transect
interval number.
}
\description{
Create echogram curtain plot from interval-layer data.
}
\details{
This function is intended to build on the plots in exploreACMT(). The plots
provide a view of the data that is not possible when looking at just Sv
or TS/sigma, as these two can interact in such a way as to generate fish
density that might be higher than expected given the appearance of the Sv
data.One key use for this sort of plot is to look for analysis cells that
have density that is higher than expected because of low numbers of targets.
}
\examples{

}
