% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateACMT.R
\name{estimateACMT}
\alias{estimateACMT}
\title{Estimate Fish Density and Biomass from Acoustic and Midwater Trawl Data}
\usage{
estimateACMT(...)
}
\arguments{
\item{...}{any arguments}
}
\description{
Estimate fish density (number per ha) and biomass (kg per ha) from
acoustic and midwater trawl data.
}
\details{
This function is now deprecated.
  Use \code{\link{estimateLake}} instead.
}
