% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMulti.R
\name{mapMulti}
\alias{mapMulti}
\title{Multipanel Map of Locations}
\usage{
mapMulti(
  bygroup,
  sug = sort(unique(bygroup)),
  plottext = FALSE,
  ID = 1,
  emphasis = NULL,
  short = TRUE,
  lon,
  lat,
  samescale = TRUE,
  rlon = NULL,
  rlat = NULL,
  cushion = 0.1,
  IDcol = NULL,
  mapcol = "gray",
  boxcol = "gray",
  misscol = "brown",
  misstext = " - No sites",
  mar = c(0, 0, 2.5, 0)
)
}
\arguments{
\item{bygroup}{Vector, identifying the group membership of the locations to be mapped.}

\item{sug}{Vector, identifying the unique groups to which locations may belong,
default sort(unique(\code{bygroup})).}

\item{plottext}{Logical scalar indicating if text (TRUE) or symbols (FALSE, default)
should be printed at each location.}

\item{ID}{Vector, plot symbol (if \code{plottext}=FALSE) or text
(if \code{plottext}=TRUE)
to print on the map at each location, either of length 1 or
the same length as \code{bygroup}, default 1.}

\item{emphasis}{Logical vector, indicating observations that should be emphasized on the
map, same length as \code{bygroup}, default NULL.}

\item{short}{Logical scalar, indicating aspect of map area.  If TRUE, the default,
the mapped area is assumed to be wider (longitudinally) than tall.
Used to better arrange multiple maps on a single page.}

\item{lon}{A numeric vector of longitudes in decimal degrees.}

\item{lat}{A numeric vector of latitudes in decimal degrees.
Same length as \code{lon}.}

\item{samescale}{Logical scalar, indicating if the same (TRUE, default) or different
(FALSE) lon/lat scales should be used for all panels.}

\item{rlon}{A numeric vector of length 2, range of longitudes to map,
in decimal degrees.  The default, NULL, means that the range of \code{lon}
will be used, either over all panels (if \code{samescale}=TRUE) or
for each panel separately (if \code{samescale}=FALSE).}

\item{rlat}{A numeric vector of length 2, range of latitudes to map,
in decimal degrees.  The default, NULL, means that the range of \code{lat}
will be used, either over all panels (if \code{samescale}=TRUE) or
for each panel separately (if \code{samescale}=FALSE).}

\item{cushion}{A numeric scalar indicating the amount of cushion to add to the \code{rlon}
and \code{rlat} ranges in decimal degrees, default 0.1.}

\item{IDcol}{A vector, the color used to map locations, same length as \code{bygroup}.
If NULL, the default, a range of colors will be assigned automatically.}

\item{mapcol}{A scalar, the color used to draw the map lines (e.g., lake boundary),
default "gray".}

\item{boxcol}{A scalar, the color used to draw the box around the map, default "gray".}

\item{misscol}{A scalar, the color used to label maps with no locations in the
given bygroup, default "brown".}

\item{misstext}{A character scalar, the text used to label maps with no locations in the
given bygroup, default " - No sites".}

\item{mar}{A numeric vector of length 4, the number of lines of margin
c(bottom, left, top, right) around each plotted map plot, default
c(0, 0, 2.5, 0).}
}
\description{
Multipanel map of locations, one map for each group.
}
\examples{
\dontrun{
 mygroup <- c(1, 1, 1, 2, 2, 2, 3, 3, 3)
 myID <- LETTERS[1:9]
 mylon <- rnorm(9, mean=-82)
 mylat <- rnorm(9, mean=45)
 mapMulti(bygroup=mygroup, plottext=TRUE, ID=myID,
   emphasis=myID \%in\% c("G", "A"), cushion=0, lon=mylon, lat=mylat)
 mapMulti(bygroup=mygroup, sug=1:4, short=FALSE, lon=mylon, lat=mylat,
   samescale=FALSE)
}

}
