% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeLen.R
\name{plotAgeLen}
\alias{plotAgeLen}
\title{Plot Age-LENGTH Key}
\usage{
plotAgeLen(m, inc = 0.3, zeroes = 4, fg = "red", bg = NA, ...)
}
\arguments{
\item{m}{A numeric matrix of counts or relative frequencies with the two dimensions
representing ages and lengths.}

\item{inc}{A numeric scalar giving the radius (in inches) of the bubble representing
the largest value in \code{m}, default 0.3.}

\item{zeroes}{A scalar giving the symbol to be used for plotting zeroes, default 4.
See the argument \code{pch} of the \code{\link{points}} function
for possible values and their interpretation.}

\item{fg}{A scalar giving the color the circles are to be drawn in.}

\item{bg}{A scalar giving the color with which the circles are to be filled.  The default,
NA, leaves the symbols unfilled.}

\item{...}{Additional arguments to \code{\link{plot}}.}
}
\description{
Plot age-length key as a bubble plot.
}
\details{
The matrix \code{m} can have ages in rows and lengths in
  columns or vice versa.  In either case, unique row values are plotted
  along the x-axis, unique column values are plotted along the y-axis.

  Values in the matrix \code{m} are represented as circles, with zero
  values represented by black dots.
}
\examples{
mymat <- matrix(c(10, 1, 0, 0, 5, 0, 0, 3, 2, 0, 0, 1), nrow=3,
 dimnames=list(age=1:3, len=seq(25, 55, 10)))
plotAgeLen(mymat, inc=1)
plotAgeLen(t(mymat), xlab="LENGTH", ylab="Age")

}
