% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntLay.R
\name{plotIntLay}
\alias{plotIntLay}
\title{Plot Acoustic Survey Data using Different Colored Symbols}
\usage{
plotIntLay(
  interval,
  layer,
  group,
  grouporder = sort(unique(group)),
  colorz,
  main = ""
)
}
\arguments{
\item{interval}{A numeric vector of intervals along the length of an acoustic transect.}

\item{layer}{A numeric vector of layers from surface to bottom along the vertical
water column of an acoustic transect, all values should be <= 0,
the same length as \code{interval}.}

\item{group}{A vector of group identifiers, the same length as \code{interval}.}

\item{grouporder}{A vector of unique group identifiers, providing the order that each group
will be plotted, the same length as \code{unique(group)}, default
\code{sort(unique(group))}.}

\item{colorz}{A vector of character or numeric colors to use,
the same length as \code{interval}.}

\item{main}{A character scalar of the main title of the plot, default "".}
}
\description{
Plot acoustic survey data, interval vs. layer, using different colored
symbols for data exploration purposes.  Place multiple group-specific plots
on one page, using the same x- and y-scales.
}
