% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiReader.r
\name{ReadPsi}
\alias{ReadPsi}
\title{Get transducer equivalent beam angle (EBA, dB) for all transducers
in the input Sv data and convert to psi 10^(EBA/10) for use in estNv
This implementation requires the user to respond to prompts, with
the number of prompts = length of list of unique transducers.}
\usage{
ReadPsi()
}
\arguments{
\item{sv}{Data frame of Sv data that is the result of readAll().}

\item{EVfolder}{Vector of base folder names in which the EV files used to generate the
data are location. Length is the same as the number of rows in \code{sv}.}

\item{EBA}{Numeric scalar of the equivalent beam angle. Must be negative and in dB.
Unknown what will happen if otherwise.}
}
\value{
A data frame with three columns (dat.source, psi, and EBA) and as many rows as
  the length of the number of unique transducers in the Sv data.
}
\description{
Get transducer equivalent beam angle (EBA, dB) for all transducers
in the input Sv data and convert to psi 10^(EBA/10) for use in estNv
This implementation requires the user to respond to prompts, with
the number of prompts = length of list of unique transducers.
}
\details{
If \code{EBA} is positive the calculated psi will be incorrect.
}
\references{
Based on a function posted by EuGENE on 02 February 2018 on stackoverflow
  \href{https://stackoverflow.com/questions/48588180/in-r-how-to-read-multiple-integers-from-user-prompt/48588287#48588287}{[link]}.
}
\seealso{
\code{\link{EstNv}}
}
