% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiReader.R
\name{ReadPsi}
\alias{ReadPsi}
\title{PsiReader}
\usage{
ReadPsi()
}
\value{
Creates a data frame with as many rows as unique transducers and the columns
'EV file folder name - frequency kHz', psi, and EBA.
}
\description{
Read EBA (equivalent beam angle, dB) provided by the user and convert
to psi.
}
\details{
Sv data set resulting from readAll().
}
